optional remove image

+++ src/api/banner/core/update.ts
@@ -15,24 +15,30 @@ export default async function update (req, res) {
 
   if (!bannerId) return res.sendStatus(404)
 
+  const banner = (await knex('banner').where({ bannerId }))[0]
+
+  if (!banner) return res.sendStatus(404)
+
   try {
     await knex('banner')
       .where({ bannerId })
       .update({
-        title,
-        description: description || null,
-        targetProgram: targetProgram || null,
-        targetCategory: targetCategory || null,
-        link: link || null,
-        expiresAt: expiresAt || null,
-        createdAt: new Date().getTime().toString()
+        title: title || banner.title || null,
+        description: description || banner.description || null,
+        targetProgram: targetProgram || banner.targetProgram || null,
+        targetCategory: targetCategory || banner.targetCategory || null,
+        link: link || banner.banner || null,
+        expiresAt: expiresAt || banner.expiresAt || null
       })
 
     if (req.file) {
       const { path: pathToImage } = req.file
 
-      const { image } = (await knex('banner').where({ bannerId }).select('image'))[0]
-      fs.unlinkSync(path.resolve(__dirname, '..', '..', '..', '..', '..', '..', 'www', 'storage', 'radio', image.split('radio/')[1]))
+      const { image } = banner
+
+      if (image) {
+        fs.unlinkSync(path.resolve(__dirname, '..', '..', '..', '..', '..', '..', 'www', 'storage', 'radio', image.split('radio/')[1]))
+      }
 
       const baseURL = 'http://dmihost.com.br/'
       const imageURL = baseURL + pathToImage.substring(pathToImage.indexOf('storage'))
